Minor refactoring of the 3d sam model
