+++ micro_sam/sam_3d_wrapper.py
@@ -1,8 +1,22 @@
+from typing import Type
+
 import torch
 import torch.nn as nn
-from typing import Type
+
 from segment_anything.modeling.image_encoder import window_partition, window_unpartition
 from segment_anything.modeling import Sam
+
+from .util import get_sam_model
+
+
+def get_3d_sam_model(device, n_classes, image_size, model_type="vit_b"):
+    predictor, sam = get_sam_model(
+        return_sam=True, model_type=model_type, device=device, num_multimask_outputs=n_classes,
+        flexible_load_checkpoint=True, image_size=image_size,
+    )
+    sam_3d = Sam3DWrapper(sam)
+    sam_3d.to(device)
+    return sam_3d
 
 
 class Sam3DWrapper(nn.Module):
