+++ micro_sam/training/semantic_sam_trainer.py
@@ -93,3 +93,17 @@ class SemanticSamTrainer(DefaultTrainer):
             self.logger.log_validation(self._iteration, metric_val, loss_val, x, y, masks)
 
         return metric_val
+
+
+class SemanticSamTrainer3D(SemanticSamTrainer):
+    def _get_model_outputs(self, batched_inputs):
+        model_input = torch.stack([inp["image"] for inp in batched_inputs]).to(self.device)
+        image_size = batched_inputs[0]["original_size"][-1]
+        batched_outputs = self.model(
+            model_input,
+            multimask_output=(self.num_classes > 1),
+            image_size=image_size
+        )
+        # masks = torch.stack([output["masks"].squeeze(0) for output in batched_outputs])
+        masks = batched_outputs["masks"]
+        return masks
